upload file added

+++ CourseSystem/src/main/java/com/project/CourseSystem/service/impl/UserServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -53,10 +54,24 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void updateAvatar(MultipartFile file) throws IOException {
+    public void updateAvatar(MultipartFile file) throws IOException{
         String folder = "/photos/";
         byte[] bytes = file.getBytes();
         Path path = Paths.get(folder + file.getOriginalFilename());
-        Files.write(path, bytes);
+        File f = new File("/photos");
+            if(!f.exists()){
+                f.mkdir();
+                Files.write(path, bytes);
+            }
+            else{
+                Files.write(path, bytes);
+            }
     }
+
+    @Override
+    public void saveAvatar(UserInfo userInfo) {
+        userInfoRepository.save(userInfo);
+    }
+
+
 }
